Revert "Rewrite ZipOS"
This reverts commit b01282e23ef0b9f68c9979d0c419fcfe64d5ca7c. Some tests
are broken. It's not clear how it'll impact metal yet. Let's revisit the
memory optimization benefits of this change again sometime soon.

+++ libc/runtime/zipos-stat.c
@@ -27,7 +27,7 @@
  * @asyncsignalsafe
  */
 int __zipos_stat(struct ZiposUri *name, struct stat *st) {
-  int cf;
+  ssize_t cf;
   struct Zipos *zipos;
   if (!(zipos = __zipos_get())) return enoexec();
   if ((cf = __zipos_find(zipos, name)) == -1) return -1;
